docs: migrate docs, rename org, and fix build helpers - Move quickstart.md from specs/ to docs/ and update all internal links   to be relative within docs/ - Replace all commerce-projects references with gitstore-dev across   docs, go.mod module path, Go imports, and gqlgen.yml scalar models - Delete stale T142-QUICKSTART-VALIDATION.md; apply all pending   corrections to docs/quickstart.md (git:// -> http://, product counts,   depth field, collection model, caveats) - Add relative-link rule to AGENTS.md / CLAUDE.md Tool Usage section - Restore missing graph helper functions (getProductsInCategoryTree,   getProductsInCollection, applyProductFilters) to fix go generate and   go build failures
