+++ api/cmd/server/main.go
@@ -14,16 +14,16 @@ import (
 
 	gqlhandler "github.com/99designs/gqlgen/graphql/handler"
 	"github.com/99designs/gqlgen/graphql/playground"
-	"github.com/commerce-projects/gitstore/api/internal/cache"
-	"github.com/commerce-projects/gitstore/api/internal/catalog"
-	"github.com/commerce-projects/gitstore/api/internal/graph"
-	"github.com/commerce-projects/gitstore/api/internal/graph/generated"
-	"github.com/commerce-projects/gitstore/api/internal/handler"
-	"github.com/commerce-projects/gitstore/api/internal/health"
-	"github.com/commerce-projects/gitstore/api/internal/loader"
-	"github.com/commerce-projects/gitstore/api/internal/logger"
-	"github.com/commerce-projects/gitstore/api/internal/middleware"
-	"github.com/commerce-projects/gitstore/api/internal/websocket"
+	"github.com/gitstore-dev/gitstore/api/internal/cache"
+	"github.com/gitstore-dev/gitstore/api/internal/catalog"
+	"github.com/gitstore-dev/gitstore/api/internal/graph"
+	"github.com/gitstore-dev/gitstore/api/internal/graph/generated"
+	"github.com/gitstore-dev/gitstore/api/internal/handler"
+	"github.com/gitstore-dev/gitstore/api/internal/health"
+	"github.com/gitstore-dev/gitstore/api/internal/loader"
+	"github.com/gitstore-dev/gitstore/api/internal/logger"
+	"github.com/gitstore-dev/gitstore/api/internal/middleware"
+	"github.com/gitstore-dev/gitstore/api/internal/websocket"
 	"go.uber.org/zap"
 )
