+++ api/internal/cache/manager.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/commerce-projects/gitstore/api/internal/catalog"
+	"github.com/gitstore-dev/gitstore/api/internal/catalog"
 	"go.uber.org/zap"
 )
