+++ api/internal/graph/collection_mutations_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/commerce-projects/gitstore/api/internal/models"
+	"github.com/gitstore-dev/gitstore/api/internal/models"
 	"github.com/go-git/go-git/v5"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
