+++ api/internal/loader/category_loader_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/commerce-projects/gitstore/api/internal/catalog"
+	"github.com/gitstore-dev/gitstore/api/internal/catalog"
 	"go.uber.org/zap"
 )
