+++ api/internal/loader/context.go
@@ -5,7 +5,7 @@ package loader
 import (
 	"context"
 
-	"github.com/commerce-projects/gitstore/api/internal/catalog"
+	"github.com/gitstore-dev/gitstore/api/internal/catalog"
 	"go.uber.org/zap"
 )
