+++ api/tests/integration/category_hierarchy_test.go
@@ -5,7 +5,7 @@ package integration
 import (
 	"testing"
 
-	"github.com/commerce-projects/gitstore/api/tests/testutil"
+	"github.com/gitstore-dev/gitstore/api/tests/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
